Order Creation added

+++ be/PodoCareWebsite/src/main/java/com/podocare/PodoCareWebsite/repo/order/OrderProductRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderProductRepo extends JpaRepository<OrderProduct, Integer> {
+public interface OrderProductRepo extends JpaRepository<OrderProduct, Long> {
 }
